Part 1 of supporting an "on a stick" configuration.
Currently, it assumes that ispConfig.py will gain the following
fields:

* "OnAStick" - True or False
* "StickVlanA" - the VLAN facing the core router
* "StickVlanB" - the VLAN facing the edge router

And set interfaceA and interfaceB to point to the trunk interface.

This patch brings functionality up to: mapping to a CPU and TC
handle - but always the same one in both directions. So upload
and download will be equal. RTT and bandwidth tracking are working.
Changes to the Python code will be needed to actually use this.